move GOID to separate file
